OBPIH-6104 Move sections/subsections into separate directories
