+++ src/js/components/productSupplier/create/subsections/AdditionalDetails.jsx
